Refactored code for better separation between init and run.
The Status class now has its own run() method so that users no longer
need to refer back to the parent Umap object when iteratively updating
an existing Status. This avoids confusion with the impact on the child
Status when the parent Umap's parameters are updated.

Also, we now support explicit specification of the number of threads
used in the parallel sections. This avoids reliance on global variables
and inadvertent nested parallelization. We also added support for custom
parallelization schemes via the UMAPPP_CUSTOM_PARALLEL macro. Note that
paralleliation of the Irlba calls has not yet been implemented.