Simplified the Status::run method to reduce arguments.
Specifically, the number of dimensions and the embedding pointer was already
passed along during initialize(), so it makes little sense to require
respecification during run(), given that they will almost certainly be the
same. So now we just store both arguments in the Status and use them in run().
We also provide methods to pull out both values for inspection if necessary.