Remove batched mode as it compromises result quality too much.
More specifically, the outcome from batch updates is too different to that of
the iterative updates, such that it's difficult to recommend the former as a
drop-in replacement for the latter. Even though it's easier to parallelize per
epoch, the convergence is such that we need to spend more epochs (briefly
discussed in jlmelville/uwot#83), which defeats the purpose.

By removing batch processing, we also simplify the code a bit, which is nice.