Parallelized the layout optimization step. (#14)
This uses a pretty wild parallelization scheme to overcome the fact that each
observation's update mutates other observations. We distribute individual
observations to each worker, stopping whenever an observation's updates affects
something that a currently-active worker is relying on.  Once all currently
active workers have finished, we then resume distribution.  This hopes that,
for large datasets, the dependency graph is relatively sparse, such that most
observations do not require a stop.

Because we're doing so much distribution of jobs to workers, the overhead of
parallelization is pretty high. We can't afford mutexes to wake threads;
instead, we use a busy-wait on each thread such that a job can be picked up
instantly. While this eliminates latency, this has the side-effects of
maintaining 100% CPU usage even if the thread is waiting. This can cause major
thread contention if the number of physical cores is less than the number of
waiting workers.

As a result of this, and the cost-inefficiency of 100% CPU usage when nothing
is happening, we don't parallelize the layout optimization by default. But if
speed is required and resources are available, users can turn it on to get a
moderate speed boost.