Constrain CPU usage for parallel testing on CI systems.
In particular, GitHub only provides 2 CPUs per job. This causes some delays due
to contention when more threads than CPUs are requested; this is especially
bad for the parallelized optimization due to its busy-waiting strategy.

At the same time, though, we want to be able to test the behavior of
parallelization with a higher number of threads, just to check that the job
distribution is done correctly. So, we set the default to 3 and we override it
on a CI-specific basis (in this case, in the GitHub actions).