Convert NeighborList back into an alias of knncolle's class.
This reduces the number of classes that a user has to learn about. It
allows use to just use knncolle's find_nearest_neighbors output
directly, which simplifies one of the initialize() overloads.

This requires the latest version of knncolle that reverts back to the v1
definition of the NeighborList, which also requires the addition of the
subpar dependency to the R package Makevars.