Use Newton's method to find the sigma for computing neighbor similarities.   (#23)
This converges much more quickly to the specified tolerance than the binary
search. We still keep the binary search code around, as a fallback in (rare?)
cases where Newton's goes off-track. This approach is based on the very similar
search in the qdtsne package for the bandwidth of the Gaussian kernel.

We do some modest refactoring to improve the efficiency of the inner
Newton/binary search iterations. We flesh out the tests for the edge cases. We
also move all the defaulted arguments into an (internal) Options class for
easier testing. This also revealed a minor bug where we didn't pass along the
specified number of threads to neighbor_similarities() - now fixed.